Modify whitespace handling and add bruteforce option
This update modifies the project's whitespace handling! Previously we
would trim all whitespace from the input. Now, spaces are ignored when
encrypting and decrypting. This is for usability and ease of reading
back decrypted message.

caesar_break now has an option to run a bruteforce attack, made
especially for when you only have short ciphertext (and thus frequency
analysis fails). Also, resolved a a boundary issue where we could
go outside the A-Z character range when iterating possible keys.